refactor updates
