+++ frontend/constants/serviceTemplates/index.ts
@@ -1,8 +1,8 @@
 export { AGENT_UI_RELEASES, type AgentUiRelease } from './agentUiReleases';
 export { KPI_DESC_PREFIX } from './constants';
 export {
+  BASIUS_SERVICE_TEMPLATE,
   MODIUS_SERVICE_TEMPLATE,
   OPTIMUS_SERVICE_TEMPLATE,
 } from './service/babydegen';
-export { BASIUS_SERVICE_TEMPLATE } from './service/basius';
 export { SERVICE_TEMPLATES } from './serviceTemplates';
